implement break-based unit interruption
It is now possible to interrupt units in their non-critical sections
when they need to be obsoleted. The break exception propagates to all
dependents, killing-off the entire dependency tree.

All units start with breaks disabled just to be on the safe side,
but any long running tasks should allow interruption and handle it
gracefully.

Signed-off-by: Jan Dvořák <mordae@anilinux.org>