improve interface of configuration objects
Add a generic method to detect when the object is going to be
destroyed and should be deleted from the configuration when it's
task finishes.

Also change how units are spawned. Instead of receiving a manager,
pass them a hash table with other units so that a snapshot can be
passed to them with every transaction.

Signed-off-by: Jan Dvořák <mordae@anilinux.org>